refactor: extract collect_once(), fix orphan cancel token, remove chrono, clean up trace.rs
- Extract shared collect_once() helper into src/commands/mod.rs to deduplicate
  the ~100-line read-collect-json loop between pull.rs and trace.rs
- Fix orphan cancel token: use cancel.child_token() instead of
  CancellationToken::new() so Ctrl+C propagates to in-flight reads
- Remove chrono dependency; use humantime::format_rfc3339() instead
- Remove std::process::exit() calls from run_trace(); only trace_command
  entry point calls process::exit (matching pull_command pattern)
- Add TODO comment for persistent connection optimization
