+++ Cargo.toml
@@ -26,7 +26,6 @@ clap = { version = "4", features = ["derive"] }
 anyhow = "1"
 indexmap = { version = "2", features = ["serde"] }
 dashmap = "6"
-chrono = "0.4"
 humantime = "2"
 humantime-serde = "1"
 async-trait = "0.1"
