+++ src/commands/mod.rs
@@ -5,8 +5,11 @@ pub mod trace;
 
 use anyhow::Result;
 use regex::Regex;
+use serde_json::json;
+use tokio_util::sync::CancellationToken;
 
 use crate::config;
+use crate::reader::{MetricReaderFactory, MetricReaderFactoryImpl};
 
 /// Filter collectors by name regex and metrics by metric regex.
 /// Returns a new list of collectors with only matching metrics.
@@ -40,4 +43,112 @@ pub fn filter_collectors(
         }
     }
     Ok(filtered)
+}
+
+/// Shared collection logic used by both `pull` and `trace`.
+///
+/// Connects to each collector, reads metrics, and returns JSON objects plus
+/// counters. Returns `(collectors_json, total_metrics, successful, failed)`.
+///
+/// TODO: Consider keeping persistent connections across iterations instead of
+/// connect/disconnect per call. The current approach is correct but slower for
+/// short polling intervals.
+pub async fn collect_once(
+    collectors: &[config::CollectorConfig],
+    cancel: &CancellationToken,
+) -> (Vec<serde_json::Value>, usize, usize, usize) {
+    let factory = MetricReaderFactoryImpl;
+    let mut total_metrics: usize = 0;
+    let mut successful: usize = 0;
+    let mut failed: usize = 0;
+    let mut collectors_json = Vec::new();
+
+    for collector in collectors {
+        let mut reader = match factory.create(collector) {
+            Ok(r) => r,
+            Err(e) => {
+                let mut metrics_json = Vec::new();
+                for metric_cfg in &collector.metrics {
+                    total_metrics += 1;
+                    failed += 1;
+                    metrics_json.push(json!({
+                        "name": metric_cfg.name,
+                        "value": null,
+                        "raw_value": null,
+                        "error": format!("collector create failed: {e}")
+                    }));
+                }
+                collectors_json.push(json!({
+                    "name": collector.name,
+                    "protocol": collector.protocol.to_string(),
+                    "metrics": metrics_json
+                }));
+                continue;
+            }
+        };
+        reader.set_metrics(collector.metrics.clone());
+        if let Err(e) = reader.connect().await {
+            let mut metrics_json = Vec::new();
+            for metric_cfg in &collector.metrics {
+                total_metrics += 1;
+                failed += 1;
+                metrics_json.push(json!({
+                    "name": metric_cfg.name,
+                    "value": null,
+                    "raw_value": null,
+                    "error": format!("connect failed: {e}")
+                }));
+            }
+            collectors_json.push(json!({
+                "name": collector.name,
+                "protocol": collector.protocol.to_string(),
+                "metrics": metrics_json
+            }));
+            continue;
+        }
+        let results = reader.read(cancel).await;
+        let _ = reader.disconnect().await;
+
+        let mut metrics_json = Vec::new();
+        for metric_cfg in &collector.metrics {
+            total_metrics += 1;
+            match results.metrics.get(&metric_cfg.name) {
+                Some(Ok((raw_value, scaled_value))) => {
+                    successful += 1;
+                    metrics_json.push(json!({
+                        "name": metric_cfg.name,
+                        "value": scaled_value,
+                        "raw_value": raw_value,
+                        "error": null
+                    }));
+                }
+                Some(Err(e)) => {
+                    failed += 1;
+                    metrics_json.push(json!({
+                        "name": metric_cfg.name,
+                        "value": null,
+                        "raw_value": null,
+                        "error": e.to_string()
+                    }));
+                }
+                None => {
+                    failed += 1;
+                    metrics_json.push(json!({
+                        "name": metric_cfg.name,
+                        "value": null,
+                        "raw_value": null,
+                        "error": "metric not in results"
+                    }));
+                }
+            }
+        }
+
+        collectors_json.push(json!({
+            "name": collector.name,
+            "protocol": collector.protocol.to_string(),
+            "metrics": metrics_json
+        }));
+    }
+
+    (collectors_json, total_metrics, successful, failed)
 }
