+++ src/commands/pull.rs
@@ -6,10 +6,8 @@ use std::path::Path;
 
 use crate::config::{find_config_file, Config};
 use crate::logging::{init_logging, map_logging_config, LogOutput, LoggingConfig};
-use crate::reader::MetricReaderFactory;
-use crate::reader::MetricReaderFactoryImpl;
 
-use super::filter_collectors;
+use super::{collect_once, filter_collectors};
 
 /// Entry point for the `pull` subcommand.
 ///
@@ -65,105 +63,9 @@ pub async fn run_pull(
         bail!("no collectors/metrics match the given filters");
     }
 
-    let factory = MetricReaderFactoryImpl;
     let cancel = CancellationToken::new();
-    let mut total_metrics: usize = 0;
-    let mut successful: usize = 0;
-    let mut failed: usize = 0;
-    let mut collectors_json = Vec::new();
-
-    for collector in &filtered_collectors {
-        let mut reader = match factory.create(collector) {
-            Ok(r) => r,
-            Err(e) => {
-                // Report all metrics as failed for this collector
-                let mut metrics_json = Vec::new();
-                for metric_cfg in &collector.metrics {
-                    total_metrics += 1;
-                    failed += 1;
-                    metrics_json.push(json!({
-                        "name": metric_cfg.name,
-                        "value": null,
-                        "raw_value": null,
-                        "error": format!("collector create failed: {e}")
-                    }));
-                }
-                let protocol_name = collector.protocol.to_string();
-                collectors_json.push(json!({
-                    "name": collector.name,
-                    "protocol": protocol_name,
-                    "metrics": metrics_json
-                }));
-                continue;
-            }
-        };
-        reader.set_metrics(collector.metrics.clone());
-        if let Err(e) = reader.connect().await {
-            // Report all metrics as failed for this collector
-            let mut metrics_json = Vec::new();
-            for metric_cfg in &collector.metrics {
-                total_metrics += 1;
-                failed += 1;
-                metrics_json.push(json!({
-                    "name": metric_cfg.name,
-                    "value": null,
-                    "raw_value": null,
-                    "error": format!("connect failed: {e}")
-                }));
-            }
-            let protocol_name = collector.protocol.to_string();
-            collectors_json.push(json!({
-                "name": collector.name,
-                "protocol": protocol_name,
-                "metrics": metrics_json
-            }));
-            continue;
-        }
-        let results = reader.read(&cancel).await;
-        let _ = reader.disconnect().await;
-
-        let mut metrics_json = Vec::new();
-        for metric_cfg in &collector.metrics {
-            total_metrics += 1;
-            match results.metrics.get(&metric_cfg.name) {
-                Some(Ok((raw_value, scaled_value))) => {
-                    successful += 1;
-                    metrics_json.push(json!({
-                        "name": metric_cfg.name,
-                        "value": scaled_value,
-                        "raw_value": raw_value,
-                        "error": null
-                    }));
-                }
-                Some(Err(e)) => {
-                    failed += 1;
-                    metrics_json.push(json!({
-                        "name": metric_cfg.name,
-                        "value": null,
-                        "raw_value": null,
-                        "error": e.to_string()
-                    }));
-                }
-                None => {
-                    failed += 1;
-                    metrics_json.push(json!({
-                        "name": metric_cfg.name,
-                        "value": null,
-                        "raw_value": null,
-                        "error": "metric not in results"
-                    }));
-                }
-            }
-        }
-
-        let protocol_name = collector.protocol.to_string();
-
-        collectors_json.push(json!({
-            "name": collector.name,
-            "protocol": protocol_name,
-            "metrics": metrics_json
-        }));
-    }
+    let (collectors_json, total_metrics, successful, failed) =
+        collect_once(&filtered_collectors, &cancel).await;
 
     let output = json!({
         "collectors": collectors_json,
