+++ src/commands/trace.rs
@@ -1,5 +1,4 @@
 use anyhow::{bail, Context, Result};
-use chrono::Utc;
 use serde_json::json;
 use std::path::Path;
 use std::time::Duration;
@@ -8,9 +7,8 @@ use tokio_util::sync::CancellationToken;
 
 use crate::config::{find_config_file, CollectorConfig, Config};
 use crate::logging::{init_logging, map_logging_config, LogOutput, LoggingConfig};
-use crate::reader::{MetricReaderFactory, MetricReaderFactoryImpl};
 
-use super::filter_collectors;
+use super::{collect_once, filter_collectors};
 
 /// Entry point for the `trace` subcommand.
 pub async fn trace_command(
@@ -92,7 +90,6 @@ async fn run_trace(
     interval: Duration,
     cancel: &CancellationToken,
 ) -> i32 {
-    let factory = MetricReaderFactoryImpl;
     let mut iteration: u64 = 0;
     let mut total_successful: usize = 0;
     let mut total_failed: usize = 0;
@@ -103,100 +100,11 @@ async fn run_trace(
         }
 
         iteration += 1;
-        let timestamp = Utc::now().to_rfc3339();
-        let inner_cancel = CancellationToken::new();
+        let timestamp = humantime::format_rfc3339(std::time::SystemTime::now()).to_string();
+        let inner_cancel = cancel.child_token();
 
-        let mut total_metrics: usize = 0;
-        let mut successful: usize = 0;
-        let mut failed: usize = 0;
-        let mut collectors_json = Vec::new();
-
-        for collector in collectors {
-            let mut reader = match factory.create(collector) {
-                Ok(r) => r,
-                Err(e) => {
-                    let mut metrics_json = Vec::new();
-                    for metric_cfg in &collector.metrics {
-                        total_metrics += 1;
-                        failed += 1;
-                        metrics_json.push(json!({
-                            "name": metric_cfg.name,
-                            "value": null,
-                            "raw_value": null,
-                            "error": format!("collector create failed: {e}")
-                        }));
-                    }
-                    collectors_json.push(json!({
-                        "name": collector.name,
-                        "protocol": collector.protocol.to_string(),
-                        "metrics": metrics_json
-                    }));
-                    continue;
-                }
-            };
-            reader.set_metrics(collector.metrics.clone());
-            if let Err(e) = reader.connect().await {
-                let mut metrics_json = Vec::new();
-                for metric_cfg in &collector.metrics {
-                    total_metrics += 1;
-                    failed += 1;
-                    metrics_json.push(json!({
-                        "name": metric_cfg.name,
-                        "value": null,
-                        "raw_value": null,
-                        "error": format!("connect failed: {e}")
-                    }));
-                }
-                collectors_json.push(json!({
-                    "name": collector.name,
-                    "protocol": collector.protocol.to_string(),
-                    "metrics": metrics_json
-                }));
-                continue;
-            }
-            let results = reader.read(&inner_cancel).await;
-            let _ = reader.disconnect().await;
-
-            let mut metrics_json = Vec::new();
-            for metric_cfg in &collector.metrics {
-                total_metrics += 1;
-                match results.metrics.get(&metric_cfg.name) {
-                    Some(Ok((raw_value, scaled_value))) => {
-                        successful += 1;
-                        metrics_json.push(json!({
-                            "name": metric_cfg.name,
-                            "value": scaled_value,
-                            "raw_value": raw_value,
-                            "error": null
-                        }));
-                    }
-                    Some(Err(e)) => {
-                        failed += 1;
-                        metrics_json.push(json!({
-                            "name": metric_cfg.name,
-                            "value": null,
-                            "raw_value": null,
-                            "error": e.to_string()
-                        }));
-                    }
-                    None => {
-                        failed += 1;
-                        metrics_json.push(json!({
-                            "name": metric_cfg.name,
-                            "value": null,
-                            "raw_value": null,
-                            "error": "metric not in results"
-                        }));
-                    }
-                }
-            }
-
-            collectors_json.push(json!({
-                "name": collector.name,
-                "protocol": collector.protocol.to_string(),
-                "metrics": metrics_json
-            }));
-        }
+        let (collectors_json, total_metrics, successful, failed) =
+            collect_once(collectors, &inner_cancel).await;
 
         total_successful += successful;
         total_failed += failed;
